Do not allow send dialog during sync
Only allow dna:// dialog at appropriate states
